Make `node:fs` load lazily
