+++ src/js/internal/fs/FSWatcher-lazy.ts
@@ -1,0 +1,22 @@
+var FSWatcher;
+
+var load = () => {
+  FSWatcher = require("internal/fs/FSWatcher").FSWatcher;
+  load = () => {};
+  return FSWatcher;
+};
+
+export default {
+  FSWatcherPropertyDescriptor: {
+    enumerable: true,
+    get() {
+      load();
+      return FSWatcher;
+    },
+    set() {},
+  },
+  watch: function watch(...args) {
+    load();
+    return new FSWatcher(...args);
+  },
+};
