+++ src/js/internal/fs/ReadStream-lazy.ts
@@ -1,0 +1,20 @@
+var ReadStream;
+function load() {
+  return require("internal/fs/ReadStream").ReadStream;
+}
+
+export default {
+  ReadStreamPropertyDescriptor: {
+    enumerable: true,
+
+    get() {
+      return (ReadStream ??= load());
+    },
+
+    set() {},
+  },
+
+  createReadStream(path, options) {
+    return new (ReadStream ??= load())(path, options);
+  },
+};
